Rename cmakedoc to cminx
