+++ README.md
@@ -1,48 +1,36 @@
-# CMakeDoc
+# cminx
 [![Build Status](https://travis-ci.com/CMakePP/CMakeDoc.svg?branch=master)](
 https://travis-ci.com/CMakePP/CMakeDoc)
 [![codecov](
 https://codecov.io/gh/CMakePP/CMakeDoc/branch/master/graph/badge.svg)](
 https://codecov.io/gh/CMakePP/CMakeDoc)
 [![Documentation Status](
-https://readthedocs.org/projects/cmakedoc/badge/?version=latest)](
-https://cmakedoc.readthedocs.io/en/latest/?badge=latest)
+https://readthedocs.org/projects/cminx/badge/?version=latest)](
+https://cminx.readthedocs.io/en/latest/?badge=latest)
 
 A common software practice is to keep documentation close to the functions and
 classes that are being documented (ideally in the same source file). Most
 languages have mechanisms for doing this for example Python has docstrings and
-C++ has Doxygen. To our knowledge no such solution exists for CMake. CMakeDoc
+C++ has Doxygen. To our knowledge no such solution exists for CMake. CMinx
 fixes this.
 
-CMakeDoc is a Python package that extracts documentation from CMake source
+CMinx is a Python package that extracts documentation from CMake source
 files. CMake developers document their functions by immediately proceeding the
-function with a comment block written in reStructuredText (reST). CMakeDoc will
+function with a comment block written in reStructuredText (reST). CMinx will
 extract the comment blocks verbatim and create `*.rst` files which can be
 included in a normal Sphinx documentation set-up.
 
 ## Installing
 You can find installation instructions here: [How to install](docs/source/installation.rst)
 
-#### Manually
-Run the following commands one at a time:
-
-```console
-foo@bar:~$ git clone github.com/CMakePP/CMakeDoc.git
-foo@bar:~$ cd CMakeDoc/
-foo@bar:~/CMakeDoc$ python3 -m venv virt-env #Create our virtual environment
-foo@bar:~/CMakeDoc$ source virt-env/bin/activate #Activate virtual environment
-(virt-env) foo@bar:~/CMakeDoc$ pip3 install wheel #Necessary dependency to install manually with pip
-(virt-env) foo@bar:~/CMakeDoc$ pip3 install . #If pip installed
-(virt-env) foo@bar:~/CMakeDoc$ python3 setup.py install #If pip not installed
-```
 
 ## Usage
 For each CMake function or variable that you would like to document, prepend it with a block doc-comment.
 A block doc-comment begins with `#[[[` and ends with `#]]`.
 
-Then run `cmakedoc` on your CMake files, outputting to a directory of your choosing. The help text is printed below for reference:
+Then run `cminx` on your CMake files, outputting to a directory of your choosing. The help text is printed below for reference:
 ```
-Usage: cmakedoc [-h] [-o OUTPUT] [-r] file [file ...]
+Usage: cminx [-h] [-o OUTPUT] [-r] file [file ...]
 
 positional arguments:
   file                  CMake file to generate documentation for. If
@@ -115,7 +103,7 @@ To generate the documentation, we enter our system shell (example assumes Bash-l
 
 Generating documentation in directory `output`:
 ```console
-foo@bar:~$ cmakedoc -o output/ example.cmake
+foo@bar:~$ cminx -o output/ example.cmake
 Writing RST files to /home/foo/output
 Writing for file /home/foo/example.cmake
 Writing RST file /home/foo/output/example.rst
