+++ docs/source/conf.py
@@ -4,7 +4,7 @@ import os
 import sys
 now = datetime.datetime.now()
 
-project = 'CMakeDoc'
+project = 'CMinx'
 author = 'Ryan M. Richard'
 version = '1.0.0'  # The short X.Y version
 release = '1.0.0alpha'  # The full version, including alpha/beta/rc tags
