--- examples/sphinx/examples/example.rst
@@ -1,47 +1,0 @@
-
-######################
-examples/example.cmake
-######################
-
-.. function:: say_hi_to(person me)
-   
-   This function has very basic documentation.
-   
-   This function's description stays close to idealized formatting and does not do
-   anything fancy.
-   
-   :param person: The person this function says hi to
-   :param me: What my name is
-   :type person: string
-   :type me: string
-   
-
-
-.. function:: macro_say_hi(person)
-
-   .. warning:: This is a macro, and so does not introduce a new scope.
-
-   
-   This macro says hi
-   
-
-
-.. data:: MyList
-   
-   list
-   
-
-   :Default value: ['"Value"', '"Value 2"']
-
-   :type: VarType.List
-
-
-.. data:: MyString
-   
-   string
-   
-
-   :Default value: String
-
-   :type: VarType.String
-
