+++ examples/sphinx/source/index.rst
@@ -1,15 +1,15 @@
-.. CMake Documentation Example documentation master file, created by
-   sphinx-quickstart on Mon Jan 27 20:07:20 2020.
+.. CMinx Examples documentation master file, created by
+   sphinx-quickstart on Tue Mar 10 20:13:09 2020.
    You can adapt this file completely to your liking, but it should at least
    contain the root `toctree` directive.
 
-Welcome to CMake Documentation Example's documentation!
-=======================================================
+Welcome to CMinx Examples's documentation!
+==========================================
 
 .. toctree::
    :maxdepth: 2
    :caption: Contents:
-
+   
    example
