+++ main.py
@@ -1,13 +1,13 @@
 #!/usr/bin/env python3
 
 """
-Convenience script to call cmakedoc.main()
+Convenience script to call cminx.main()
 
 :Author: Branden Butler
 :License: Apache 2.0
 """
 
-from cmakedoc import *
+from cminx import *
 
 if __name__ == "__main__":
      main()
