+++ tests/.coveragerc
@@ -1,2 +1,2 @@
 [run]
-include=../cmakedoc/*
+include=../cminx/*
