+++ tests/context.py
@@ -2,5 +2,5 @@ import os
 import sys
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
 
-import cmakedoc
+import cminx
