+++ tests/test_documenter.py
@@ -3,9 +3,9 @@ import context
 import unittest
 import sys
 import docutils
-from cmakedoc.rstwriter import Directive
-from cmakedoc.parser.aggregator import FunctionDocumentation, MacroDocumentation, VariableDocumentation
-from cmakedoc.documenter import Documenter
+from cminx.rstwriter import Directive
+from cminx.parser.aggregator import FunctionDocumentation, MacroDocumentation, VariableDocumentation
+from cminx.documenter import Documenter
 from rst_validator import RSTValidator
